Ing : feat/board
board s3 upload file 삭제 기능 구현 중.
- url로 파일 조회가 안됨

+++ src/main/java/com/example/vvs/domain/common/s3/S3Service.java
@@ -1,4 +1,4 @@
-package com.example.vvs.domain.common;
+package com.example.vvs.domain.common.s3;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
@@ -6,6 +6,8 @@ import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.example.vvs.domain.board.dto.BoardRequestDTO;
 import com.example.vvs.domain.member.entity.Member;
+import com.example.vvs.exception.ApiException;
+import com.example.vvs.exception.ErrorHandling;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -20,6 +22,9 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
+
+import static com.example.vvs.exception.ErrorHandling.FILE_NOT_FOUND;
+
 
 @Slf4j
 @Component
@@ -45,6 +50,7 @@ public class S3Service {
         }
     }
 
+    // S3에 업로드하기위해선 MultipartFile을 File객체로 변환
     private Optional<File> convert(MultipartFile multipartFile) throws IOException {
 
         File convertFile = new File(Objects.requireNonNull(multipartFile.getOriginalFilename()));
@@ -59,15 +65,13 @@ public class S3Service {
         return Optional.empty();
     }
 
-    private String boardUpload(File uploadFile) throws IOException {
-
+    private void boardUpload(File uploadFile) throws IOException {
         String fileName = "board" + "/" + UUID.randomUUID();
-        uploadImageUrl  += uploadImageUrl.equals("") ? putS3(uploadFile, fileName) : ", " + putS3(uploadFile, fileName) ;
-        removeNewFile(uploadFile);
-        return uploadImageUrl;
+        uploadImageUrl  += uploadImageUrl.equals("") ? putS3(uploadFile, fileName) : ", " + putS3(uploadFile, fileName); // 링크 ","로 이어 붙혀서 하나의 String으로 저장.
+        removeNewFile(uploadFile); // putS3에서 AWS에 올릴 때 로컬에 파일생성 후 업로드 하게 되는데, 업로드 후 로컬파일이 필요없으니 이를 제거.
     }
 
-    private String putS3(File uploadFile, String fileName) {
+    private String putS3(File uploadFile, String fileName) { // S3에 파일 업로드하고 url를 리턴하는 핵심코드!
 
         S3Client.putObject(new PutObjectRequest(bucketName, fileName, uploadFile)
                 .withCannedAcl(CannedAccessControlList.PublicRead));
@@ -87,7 +91,11 @@ public class S3Service {
         return uploadImageUrl;
     }
 
+
     public void deleteFile(String fileName) {
+        if (!S3Client.doesObjectExist(bucketName, fileName)) {
+            throw new ApiException(FILE_NOT_FOUND);
+        }
 
         S3Client.deleteObject(new DeleteObjectRequest(bucketName, fileName));
     }
